backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
setuptools<81
pymongo==4.16.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
python-dotenv==1.2.2
requests==2.33.1
email-validator==2.3.0
razorpay==1.4.2
PyMySQL==1.1.1
